Add additional param offset
Param offset is needed when you are working with fixed elements. It fix the problems when you have some header which is fixed at the top of the page, and you are trying to check if some item is visible. When item is under the fixed header block, user doesn't see this item, but plugin returns true (item is visible), since it top offset is greater than 0. If you Pass param offset to visible function you'll check whether element is visible under offset. Param offset accepts simple number or object {top: 1, bottom: 2} etc.